chore: refactor project structure
To resemble a CLI app more, move the cobra-specific parts into cmd, and
implementation under internal.
